frontend: Fix scanning of Campaign

+++ cmd/frontend/types/types.go
@@ -12,9 +12,9 @@ type Campaign struct {
 	ID              int64
 	Name            string
 	Description     string
-	AuthorID        int32
-	NamespaceUserID int32
-	NamespaceOrgID  int32
+	AuthorID        int64
+	NamespaceUserID int64
+	NamespaceOrgID  int64
 	CreatedAt       time.Time
 	UpdatedAt       time.Time
 }
